api: reduce dependencies when used by client
When the nyuds-api crate is used by client, it only cares about data
types defined by the API, so we should get rid of extra/unnecessary
dependencies.
