+++ api/Cargo.toml
@@ -9,17 +9,20 @@ repository = "https://github.com/dragonflyoss/image-service"
 edition = "2018"
 
 [dependencies]
-dbs-uhttp = { version = "0.3.0" }
-http = "0.2.1"
-lazy_static = "1.4.0"
+dbs-uhttp = { version = "0.3.0", optional = true }
+http = { version = "0.2.1", optional = true }
+lazy_static = { version = "1.4.0", optional = true }
 libc = "0.2"
 log = "0.4.8"
-mio = { version = "0.8", features = ["os-poll", "os-ext"]}
-serde = { version = "1.0.110", features = ["rc"] }
-serde_derive = "1.0.110"
+mio = { version = "0.8", features = ["os-poll", "os-ext"], optional = true }
+serde = { version = "1.0.110", features = ["rc", "serde_derive"] }
 serde_json = "1.0.53"
-url = "2.1.1"
-vmm-sys-util = "0.10"
+url = { version = "2.1.1", optional = true }
 
 nydus-error = { version = "0.2", path = "../error" }
-nydus-utils = { version = "0.3", path = "../utils" }
+
+[dev-dependencies]
+vmm-sys-util = { version = "0.10" }
+
+[features]
+handler = ["dbs-uhttp", "http", "lazy_static", "mio", "url"]
