+++ api/src/http_endpoint_v2.rs
@@ -9,9 +9,11 @@
 use dbs_uhttp::{Method, Request, Response};
 
 use crate::http::{
+    ApiError, ApiRequest, ApiResponse, ApiResponsePayload, BlobCacheObjectId, HttpError,
+};
+use crate::http_handler::{
     error_response, extract_query_part, parse_body, success_response, translate_status_code,
-    ApiError, ApiRequest, ApiResponse, ApiResponsePayload, BlobCacheObjectId, EndpointHandler,
-    HttpError, HttpResult,
+    EndpointHandler, HttpResult,
 };
 
 /// HTTP URI prefix for API v2.
